requirements.txt
# Core dependencies
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Language detection and translation
langdetect>=1.0.9
deep-translator>=1.11.4

# Sentiment analysis
textblob>=0.17.1

# Database
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0